+++ bootloader/Disk/IDEController.h
@@ -17,7 +17,7 @@ public:
     ~IDEController();
     void initialize(PCI::Device* device,PCI::Access* access);
     bool probe_port_connected(bool is_primary,bool is_slave);
-    void read(bool is_primary,bool is_slave,uint32_t lbal,uint32_t lbah,uint16_t bytesOffset,uint16_t* buf,uint16_t bytesCount);
+    void read(bool is_primary,bool is_slave,uint32_t lbal,uint32_t lbah,uint32_t bytesOffset,uint16_t* buf,uint16_t bytesCount);
     uint16_t get_logical_sector_size(bool is_primary,bool is_slave);
     uint16_t get_physical_sector_size(bool is_primary,bool is_slave);
     uint8_t get_physical_logical_sector_alignment(bool is_primary,bool is_slave);
